requirements.txt
flask==0.12
requests==2.17.1
newsapi-python==0.2.3
